backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
autopep8==2.0.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
databases==0.7.0
dj-database-url==1.0.0
Django==4.1.3
django-heroku==0.3.1
django-nextjs==2.2.1
ecdsa==0.18.0
fastapi==0.88.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.1
multidict==6.0.2
passlib==1.7.4
psycopg2==2.9.5
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.4
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.45
sqlparse==0.4.2
starlette==0.22.0
tomli==2.0.1
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
whitenoise==6.2.0
yarl==1.8.1
